LPS-193593 Move ImageSelectorProcessor into blogs-service prepping to move ImageMagick next
