+++ modules/apps/blogs/blogs-service/src/main/java/com/liferay/blogs/internal/image/ImageSelectorProcessor.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.portal.kernel.servlet.taglib.ui;
+package com.liferay.blogs.internal.image;
 
 import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.image.ImageBag;
